fix: install

+++ install.sh
@@ -2,14 +2,10 @@
 
 set -e
 
-command -v pyenv || "Please install pyenv"  
-command -v python3 || "Prease install Python3"
+command -v python3 >/dev/null || { echo "Please install Python3"; exit 1; }
+python3 -m pip show maturin >/dev/null 2>&1 || { echo "Please install maturin (python3 -m pip install maturin)"; exit 1; }
 
-pip install . 
-cd lib
-maturin build
-pip install target/wheels/*.whl
-cd -
+python3 -m pip install . --no-build-isolation --force-reinstall
 rm -rf build lib/target src/cuckooget.egg-info cuckooget.egg-info
 
 echo "Done! How to use `ck -h`"
